refactor(ui): split pure view-model helpers out of actionNode.tsx
Move paramDisplay, leadParts, callTag, groupChildren (+ Lead/GroupItem/ChildGroup
types) into actionNode.helpers.ts. actionNode.tsx is now components-only; the
pure, JSX-free logic is isolated and unit-testable.
